+++ app/controllers/ahoy_captain/filters/pages/entry_pages_controller.rb
@@ -1,10 +1,10 @@
 module AhoyCaptain
   module Filters
     module Pages
-      # TODO: ACCOMODATE EXIT_PAGES
       class EntryPagesController < BaseController
         def index
-          query = event_query.all.with_url.distinct_url
+          query = event_query.all.distinct("entry_pages.url").select("entry_pages.url as url")
+
           render json: query.map { |row| { text: row.url } }
         end