+++ app/controllers/ahoy_captain/filters/pages/exit_pages_controller.rb
@@ -1,10 +1,9 @@
 module AhoyCaptain
   module Filters
     module Pages
-      # TODO: ACCOMODATE ENTRY_PAGES
       class ExitPagesController < BaseController
         def index
-          query = event_query.with_url.distinct_url
+          query = event_query.distinct("exit_pages.url").select("exit_pages.url as url")
 
           render json: query.map { |row| { text: row.url } }
         end