+++ test/dummy/app/models/ahoy/event.rb
@@ -6,6 +6,4 @@ class Ahoy::Event < ApplicationRecord
 
   belongs_to :visit, optional: true
   belongs_to :user, optional: true
-
-
 end